Allow running the script twice

+++ src/2023-17-1_template_upgrade.sh
@@ -15,41 +15,48 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# pull files:
-URL=https://raw.githubusercontent.com/ghga-de/microservice-repository-template
-OLD=0e1c3da09bc3eaad673dceabe69b9a8756d54eaa
-wget $URL/$OLD/.mandatory_files -O .mandatory_template_old
-wget $URL/$OLD/.static_files -O .static_template_old
-wget $URL/main/.static_files_ignore
-wget $URL/main/.mandatory_files_ignore
 
-# identify exceptions to the static and mandatory file lists:
-comm -23 \
-    <( sort .static_template_old | uniq --unique) \
-    <( sort .static_files | uniq --unique) \
-    | grep -v .mandatory_files \
-    | grep -v .static_files \
-    >> .static_files_ignore
-comm -23 \
-    <( sort .mandatory_template_old | uniq --unique) \
-    <( sort .mandatory_files | uniq --unique) \
-    | grep -v .mandatory_files \
-    | grep -v .static_files \
-    >> .mandatory_files_ignore
+if [ ! -f ./scripts/update_template_files.py ]; then
 
-# cleanup:
-rm .*_template_old
+    echo "Upgrading to new update_template_files script..."
 
-# ensure that static and mandatory files are in static list:
-echo -e ".mandatory_files\n.static_files" >> .static_files
+    # pull files:
+    URL=https://raw.githubusercontent.com/ghga-de/microservice-repository-template
+    OLD=0e1c3da09bc3eaad673dceabe69b9a8756d54eaa
+    wget $URL/$OLD/.mandatory_files -O .mandatory_template_old
+    wget $URL/$OLD/.static_files -O .static_template_old
+    wget $URL/main/.static_files_ignore
+    wget $URL/main/.mandatory_files_ignore
 
-# run updates:
-./scripts/update_static_files.py
-./scripts/update_static_files.py
-chmod +x ./scripts/update_template_files.py
-./scripts/update_template_files.py
+    # identify exceptions to the static and mandatory file lists:
+    comm -23 \
+        <( sort .static_template_old | uniq --unique) \
+        <( sort .static_files | uniq --unique) \
+        | grep -v .mandatory_files \
+        | grep -v .static_files \
+        >> .static_files_ignore
+    comm -23 \
+        <( sort .mandatory_template_old | uniq --unique) \
+        <( sort .mandatory_files | uniq --unique) \
+        | grep -v .mandatory_files \
+        | grep -v .static_files \
+        >> .mandatory_files_ignore
 
-# update license headers:
+    # cleanup:
+    rm .*_template_old
+
+    # ensure that static and mandatory files are in static list:
+    echo -e ".mandatory_files\n.static_files" >> .static_files
+
+    # run updates:
+    ./scripts/update_static_files.py
+    ./scripts/update_static_files.py
+    chmod +x ./scripts/update_template_files.py
+    ./scripts/update_template_files.py
+
+fi
+
+echo "Updating license headers..."
 find . -type f -exec \
     sed -i -e \
         's/Copyright 2021 Universität/Copyright 2021 - 2023 Universität/g' \
@@ -59,7 +66,7 @@ find . -type f -exec \
         's/Copyright 2021 - 2022 Universität/Copyright 2021 - 2023 Universität/g' \
     {} \;
 
-# update docker-in-docker feature:
+echo "Updating docker-in-docker feature..."
 perl -i -pe \
   'BEGIN{undef $/;} s|
 		"docker-in-docker": \{[^}]*\}|
